aider: Refactored `Tooba.RDF.Store` module to use the XDG data directory for storing the RDF graph and added a function to ensure the directory exists before writing the graph to the file.
# Aider chat conversation:

ASSISTANT: To store the RDF graph in a system-specific application data directory according to the XDG Base Directory Specification, we can use the `XDG` Elixir library, which provides access to the XDG directories. If you don't already have this library as a dependency, you will need to add it to your `mix.exs` file.

Here's how you can modify the `Tooba.RDF.Store` module to use the XDG data directory for storing the RDF graph:

First, ensure that the `XDG` library is added to your dependencies in `mix.exs` (you will need to add this to the chat if it's not already present):

```elixir
defp deps do
  [
    {:xdg, "~> 0.1"}
    # ... other deps
  ]
end
```

Then, run `mix deps.get` to fetch the new dependency.

Next, modify the `Tooba.RDF.Store` module to use the XDG data directory:

```elixir
lib/tooba/rdf.ex
<<<<<<< SEARCH
  @rdf_store_file_path "priv/data/rdf_store.ttl"
=======
  @app_name :tooba
  @graph_file_name "graph.ttl"

  defp rdf_store_file_path do
    xdg_data_home = XDG.Data.home!()
    Path.join([xdg_data_home, @app_name, @graph_file_name])
  end

  @rdf_store_file_path rdf_store_file_path()
>>>>>>> REPLACE
```

This change will compute the file path using the XDG data directory for the current user, creating a `tooba` subdirectory within it, and then use `graph.ttl` as the file name.

Please note that you will need to handle the creation of the `tooba` directory within the XDG data directory if it does not already exist. You can do this by adding a function to check and create the directory if necessary:

```elixir
lib/tooba/rdf.ex
<<<<<<< SEARCH
  def persist do
    graph = retrieve_graph()
    file_path = @rdf_store_file_path
=======
  defp ensure_data_dir_exists do
    file_path = rdf_store_file_path()
    dir_path = Path.dirname(file_path)

    case File.mkdir_p(dir_path) do
      :ok -> :ok
      {:error, :eexist} -> :ok
      {:error, reason} -> {:error, reason}
    end
  end

  def persist do
    :ok = ensure_data_dir_exists()
    graph = retrieve_graph()
    file_path = rdf_store_file_path()
>>>>>>> REPLACE
```

This will ensure that the directory exists before attempting to write the RDF graph to the file.

Remember to add the `XDG` library to your dependencies and fetch it before making these changes. If you need to add the `mix.exs` file to the chat for modification, please let me know.