aider: Modified `save_blob/2` to `save_blob/1` and added a helper function `generate_content_hash/1` to compute the SHA256 hash of the data and use it as the file name.
# Aider chat conversation:

USER: yeah but saving blobs should generate file names from the content hash

ASSISTANT: To modify the `Tooba.Blob.Store` module to generate file names based on the content hash, we will need to add a function that computes the hash of the given data and use that hash as the file name. We will use the `:crypto` module to compute a SHA256 hash of the data.

Here's how we can modify the `lib/tooba/blob.ex` file:

```elixir
lib/tooba/blob.ex
<<<<<<< SEARCH
  # Saves a blob to the file system.
  def save_blob(file_name, data) do
    ensure_data_dir_exists()
    file_path = blob_store_file_path(file_name)

    File.write(file_path, data)
  end
=======
  # Saves a blob to the file system with a file name based on the content hash.
  def save_blob(data) do
    ensure_data_dir_exists()
    file_name = generate_content_hash(data)
    file_path = blob_store_file_path(file_name)

    File.write(file_path, data)
  end

  # Generates a SHA256 hash of the given data.
  defp generate_content_hash(data) do
    :crypto.hash(:sha256, data)
    |> Base.encode16(case: :lower)
  end
>>>>>>> REPLACE
```

In this change, we've modified the `save_blob/2` function to `save_blob/1`, which now only takes the data as an argument and generates the file name by hashing the data. We've also added a private helper function `generate_content_hash/1` that computes the SHA256 hash of the data and encodes it as a hexadecimal string.

Please note that the `get_blob/1` function will also need to be updated to work with the new naming scheme based on content hashes. If you want to proceed with that change as well, please let me know, and I will provide the necessary modifications.