aider: Modified `save_blob/1` function to return the hash of the saved blob.
# Aider chat conversation:

USER: saving should return the hash

ASSISTANT: To modify the `save_blob/1` function in the `Tooba.Blob.Store` module to return the hash of the saved blob, we need to update the `lib/tooba/blob.ex` file. Here's the change that needs to be made:

```elixir
lib/tooba/blob.ex
<<<<<<< SEARCH
  # Saves a blob to the file system with a file name based on the content hash.
  def save_blob(data) do
    ensure_data_dir_exists()
    file_name = generate_content_hash(data)
    file_path = blob_store_file_path(file_name)

    File.write(file_path, data)
  end
=======
  # Saves a blob to the file system with a file name based on the content hash and returns the hash.
  def save_blob(data) do
    ensure_data_dir_exists()
    file_name = generate_content_hash(data)
    file_path = blob_store_file_path(file_name)

    case File.write(file_path, data) do
      :ok -> {:ok, file_name}
      {:error, reason} -> {:error, reason}
    end
  end
>>>>>>> REPLACE
```

This change will make the `save_blob/1` function return a tuple with `:ok` and the hash if the file is written successfully, or an error tuple if the file write fails.