aider: Convert `subject`, `predicate`, and `object` to strings when they are instances of `RDF.IRI` and update the template and helper function accordingly.
# Aider chat conversation:

ASSISTANT: To address the issue with `RDF.IRI` not being `Phoenix.HTML.Safe`, we need to ensure that the `IRI` values are properly converted to strings that can be safely rendered in HTML. This can be done by calling the `to_string/1` function on the `IRI` values before they are used in the HTML template.

Let's update the `lib/tooba_web/live/resource_live/index.ex` file to convert `subject`, `predicate`, and `object` to strings when they are instances of `RDF.IRI`.

Here's how you can make the changes:

```elixir
lib/tooba_web/live/resource_live/index.ex
<<<<<<< SEARCH
              <%= for {subject, predicate, object} <- RDF.Description.triples(description) do %>
                <tr>
                  <td><%= subject %></td>
                  <td><%= predicate %></td>
                  <td><%= object %></td>
                </tr>
              <% end %>
=======
              <%= for {subject, predicate, object} <- RDF.Description.triples(description) do %>
                <tr>
                  <td><%= render_iri(subject) %></td>
                  <td><%= render_iri(predicate) %></td>
                  <td><%= render_iri(object) %></td>
                </tr>
              <% end %>
>>>>>>> REPLACE
```

And then, add a helper function `render_iri/1` to the same module to handle the conversion:

```elixir
lib/tooba_web/live/resource_live/index.ex
<<<<<<< SEARCH
  defp dom_id(description) do
    RDF.IRI.to_string(description.subject) |> :erlang.phash2() |> Integer.to_string()
  end
=======
  defp dom_id(description) do
    RDF.IRI.to_string(description.subject) |> :erlang.phash2() |> Integer.to_string()
  end

  defp render_iri(iri) when is_struct(iri, RDF.IRI) do
    RDF.IRI.to_string(iri)
  end
  defp render_iri(value), do: value
>>>>>>> REPLACE
```

This change introduces a new helper function `render_iri/1` that checks if the value is an `RDF.IRI` struct and converts it to a string if it is. Otherwise, it returns the value as is. This function is then used in the template to ensure that `subject`, `predicate`, and `object` are rendered as strings that are safe for HTML.